Test/requirements.txt
google-genai
opencv-python
pillow
mss
pyaudio
pyaudio
